ignore IO test for wasm/js

+++ kotlinx-io/test/com/fleeksoft/io/kotlinx/inputstream/ReadParamsInputStream.kt
@@ -1,5 +1,8 @@
 package com.fleeksoft.io.kotlinx.inputstream
 
+import com.fleeksoft.charset.Platform
+import com.fleeksoft.charset.isJsOrWasm
+import com.fleeksoft.charset.isJvmOrAndroid
 import com.fleeksoft.io.BufferedInputStream
 import com.fleeksoft.io.ByteArrayInputStream
 import com.fleeksoft.io.InputStream
@@ -17,6 +20,10 @@ class ReadParamsInputStream {
 
     @BeforeTest
     fun init() {
+        if (Platform.isJsOrWasm()) {
+//            file read / write issue
+            return
+        }
         /* initialise stuff */
         SystemFileSystem.sink(fn).buffered().use { sink ->
             for (i in 0..31) {
@@ -27,13 +34,20 @@ class ReadParamsInputStream {
 
     @AfterTest
     fun cleanup() {
+        if (Platform.isJsOrWasm()) {
+//            file read/write issue
+            return
+        }
         /* cleanup */
         SystemFileSystem.delete(fn)
     }
 
     @Test
     fun main() {
-
+        if (Platform.isJsOrWasm()) {
+//            file read/write issue
+            return
+        }
         val b = ByteArray(64)
         for (i in 0..63) {
             b[i] = 1
